% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.POSSUM.R
\name{gen.POSSUM}
\alias{gen.POSSUM}
\title{A function to compute POSSUM scores}
\usage{
gen.POSSUM(x)
}
\arguments{
\item{x}{A dataframe or tbl where each row is a patient observation, and the columns are POSSUM predictor variables.

  x must contain the following column names (not necessarily in order):
  \describe{
  \item{Age}{continuous variable, numeric or integer}
  \item{JVP}{binary variable, whether the patient has raised JVP or not}
  \item{Cardiomegaly}{binary variable, whether the patient has cardiomegaly on CXR or not}
  \item{Oedema}{binary variable, whether the patient has peripheral oedema or not}
  \item{Warfarin}{binary variable, whether the patient normally takes warfarin or not}
  \item{Diuretics}{binary variable, whether the patient normally takes a diuretic medication or not}
  \item{AntiAnginals}{binary variable, whether the patient normally takes anti-anginal medication or not}
  \item{Digoxin}{binary variable, whether the patient normally takes digoxin or not}
  \item{AntiHypertensives}{binary variable, whether the patient normally takes blood pressure meds or not}
  \item{Dyspnoea}{categorical variable, can be: "Non" = None; "OME" = On exertion; "L" = Limiting activities; "AR" = At rest}
  \item{Consolidation}{binary variable, whether the patient has consolidation on CXR}
  \item{PulmonaryFibrosis}{binary variable, whether the patient has a history of pulmonary fibrosis or imaging findings of fibrosis}
  \item{COPD}{binary variable, whether the patient has COPD or not}
  \item{SysBP}{continuous variable, pre-op systolic blood pressure (in mmHg)}
  \item{HR}{continuous variable, pre-op pulse/heart rate (in beats per min)}
  \item{GCS}{continuous variable, pre-op Glasgow Coma Scale (3-15)}
  \item{Hb}{continuous variable, pre-op Haemoglobin (in g/L), please note the units!}
  \item{WCC}{continuous variable, pre-op White Cell Count (in * 10^9cells/L)}
  \item{Ur}{continuous variable, pre-op Urea (in mmol/L)}
  \item{Na}{continuous variable, pre-op Sodium concentration (in mmol/L)}
  \item{K}{continuous variable, pre-op Potassium concentration (in mmol/L)}
  \item{ECG}{categorical variable, can be "ND" = Not done; "NOR" = Normal ECG; "AF6090" = AF 60-90; "AF>90" = AF>90; "QW" = Q-waves; "4E" = >4 ectopics; "ST" = ST or T wave changes; "O" = Any other abnormal rhythm}
  \item{OpSeverity}{categorical variable, the surgical severity, can be Min = Minor; Int = Intermediate; Maj = Major; Xma = Xmajor; Com = Complex}
  \item{ProcedureCount}{categorical variable, number of procedures patient underwent in the last 30 days including this one, can be "1" = 1; "2" = 2; "GT2" = >2}
  \item{EBL}{categorical variable, the estimated blood loss, can be "0" = 0-100ml; "101" = 101-500ml; "501" = 501-999ml; "1000" = >=1000}
  \item{PeritonealContamination}{categorical variable, whether there was peritoneal soiling, can be "NA" = Not applicable; "NS" = No soiling; "MS" = Minor soiling; "LP" = Local pus; "FBC" = Free bowel content pus or blood}
  \item{Malignancy}{categorical variable, whether the patient has malignant disease, can be "NM" = Not malignant; "PM" = Primary malignancy only; "MNM" = Malignancy + nodal metastases; "MDM" = Malignancy + distal metastases}
  \item{OpUrgency}{categorical variable, NCEPOD classifications of urgency, can be "Ele" = Elective; "Exp" = Expedited; "U" = Urgent; "I" = Immediate}
  }}
}
\value{
A dataframe (or tbl), which you can assign to an object, with the following variables:
  \describe{
  \item{PhysScore}{The physiological score for POSSUM}
  \item{OpScore}{The operative score for POSSUM}
  \item{POSSUMLogit}{The log-odds for morbidity as calculated by POSSUM}
  \item{pPOSSUMLogit}{The log-odds for mortatlity as calculated by pPOSSUM}
  }
}
\description{
This function will parse a dataframe and produce POSSUM scores to predict perioperative mortality and morbidity. To use the function, you will need to manipulate your dataframe to have columns with the structure detailed below.
}
\section{Converting to probability scale}{

  The function will produce POSSUMLogit and pPOSSUMLogit values which are on the log-odds scale
  To convert to probabilities (0 to 1 scale), use \code{arm::invlogit()}. See: \code{\link[arm]{invlogit}}.
}

\section{References}{

  \itemize{
  \item Copeland GP, Jones D, Walters M. POSSUM: A scoring system for surgical audit. Br J Surg. 1991 Mar 1;78(3):355–60. \url{http://onlinelibrary.wiley.com/doi/10.1002/bjs.1800780327/abstract}.
  \item Prytherch DR, Whiteley MS, Higgins B, Weaver PC, Prout WG, Powell SJ. POSSUM and Portsmouth POSSUM for predicting mortality. Br J Surg. 1998 Sep 1;85(9):1217–20. \url{http://onlinelibrary.wiley.com/doi/10.1046/j.1365-2168.1998.00840.x/abstract}
  }
}

\examples{
\dontrun{
#Example of pre-processing to rename data variables to match expected column names
library(tidyverse)

test_data <- raw_data \%>\%
select(Age = S01Age,
        JVP = S03ElevatedJugularVenousPressureJvp,
        Cardiomegaly = S03RadiologicalFindingsCardiomegaly,
        Oedema = S03PeripheralOedema,
        Warfarin = S03DrugTreatmentWarfarin,
        Diuretics = S03DrugTreatmentDiureticTreatment,
        AntiAnginals = S03DrugTreatmentAntiAnginal,
        Digoxin = S03DrugTreatmentDigoxinTherapy,
        AntiHypertensives = S03DrugTreatmentAntiHypertensive,
        Dyspnoea = S03Dyspnoea,
        Consolidation = S03RadiologicalFindingsConsolidation,
        PulmonaryFibrosis = S03PastMedicalHistoryPulmonaryFibrosis,
        COPD = S03PastMedicalHistoryCOPD,
        SysBP = S03SystolicBloodPressureBpAtPreAssessment,
        HR = S03PulseRateAtPreoperativeAssessment,
        GCS = S03GlasgowComaScaleGcsPreInductionOfAnaesthesia,
        Hb = S03Hb,
        WCC = S03WhiteCellCountWcc,
        Ur = S03Urea,
        Na = S03Na,
        K = S03K,
        ECG = S03EcgFindings,
        OpSeverity = S02PlannedProcSeverity,
        ProcedureCount = S04ProcedureCount,
        EBL = S04BloodLoss,
        PeritonealContamination = S04PeritonealContamination,
        Malignancy = S04Malignancy,
        OpUrgency = S02OperativeUrgency)
}

test_data <- patients
test_output <- gen.POSSUM(test_data)
head(test_output)
}
