% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.SORT.R
\name{gen.SORT}
\alias{gen.SORT}
\title{A function to compute SORT scores}
\usage{
gen.SORT(x)
}
\arguments{
\item{x}{A dataframe or tbl where each row is a patient observation, and the columns are SORT predictor variables.

  x must contain the following column names (not necessarily in order):
  \describe{
  \item{Age}{continuous variable, numeric or integer}
  \item{ASA}{categorical variable, can be "I"; "II"; "III"; "IV"; "V"}
  \item{OpUrgency}{categorical variable, NCEPOD classifications, can be "Ele" = Elective; "Exp" = Expedited; "U" = Urgent; "I" = Immediate}
  \item{Specialty}{categorical variable, can be "Ortho" = Orthopaedics; "Colorectal" = Colorectal; "UpperGI" = Upper Gastrointestinal; "Vasc" = Vascular; "Bariatric" = Bariatric; "Other" = All else}
  \item{OpSeverity}{categorical variable, can be "Min" = Minor; "Int" = Intermediate; "Maj" = Major; "Xma" = Xmajor; "Com" = Complex}
  \item{Malignancy}{categorical variable, can be "NM" = Not malignant; "PM" = Primary malignancy only; "MNM" = Malignancy + nodal metastases; "MDM" = Malignancy + distal metastases; Or "Y" / "N"; Or "1" / "0"}
  }}
}
\value{
A dataframe (or tbl), which you can assign to an object, with the following variables:
  \describe{
  \item{SORT_mortLogit}{The log-odds for mortality as calculated by SORT}
  \item{SORT_morbLogit}{The log-odds for morbidity as calculated by SORT}
  }
}
\description{
This function will parse a dataframe and compute morbidity and mortality based on the Surgical Outcome Risk Tool (SORT). To use the function, you will need to manipulate your dataframe to have columns with the structure detailed below.
}
\section{Converting to probability scale}{

  The function will produce SORT_mortLogit and SORT_morbLogit values which are on the log-odds scale
  To convert to probabilities (0 to 1 scale), use \code{arm::invlogit()}. See: \code{\link[arm]{invlogit}}.
}

\section{References}{

  \itemize{
  \item Protopapa KL, Simpson JC, Smith NCE, Moonesinghe SR. Development and validation of the Surgical Outcome Risk Tool (SORT). Br J Surg. 2014 Dec;101(13):1774–83. \url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4240514/}.
  \item Wong DJN, Oliver CM, Moonesinghe SR. Predicting postoperative morbidity in adult elective surgical patients using the Surgical Outcome Risk Tool (SORT). BJA: British Journal of Anaesthesia. 2017;119(1):95–105. \url{https://doi.org/10.1093/bja/aex117}
  }
}

\examples{
\dontrun{
#Example of pre-processing to rename data variables to match expected column names
library(tidyverse)

test_data <- raw_data \%>\%
select(Age = S01Age,
        ASA = S03AsaPsClass,
        OpUrgency = S02OperativeUrgency,
        Specialty = Specialty,
        OpSeverity = S02PlannedProcSeverity,
        Malignancy = S04Malignancy)
}

test_data <- patients
test_output <- gen.SORT(test_data)
head(test_output)
}
