% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.SRS.R
\name{gen.SRS}
\alias{gen.SRS}
\title{A function to compute SRS scores}
\usage{
gen.SRS(x)
}
\arguments{
\item{x}{A dataframe or tbl where each row is a patient observation, and the columns are SRS predictor variables.

  x must contain the following column names (not necessarily in order):
  \describe{
  \item{ASA}{categorical variable, can be "I"; "II"; "III"; "IV"; "V"}
  \item{OpUrgency}{categorical variable, NCEPOD classifications, can be "Ele" = Elective; "Exp" = Expedited; "U" = Urgent; "I" = Immediate}
  \item{OpSeverity}{categorical variable, can be "Min" = Minor; "Int" = Intermediate; "Maj" = Major; "Xma" = Xmajor; "Com" = Complex}
  }}
}
\value{
A dataframe (or tbl), which you can assign to an object, with the following variables:
  \describe{
  \item{SRS}{The SRS score}
  \item{SRS_mortLogit}{The log-odds for mortality as calculated according to the equation \eqn{SRS_mortLogit = 0.84 * SRS - 9.81}. See reference below, page 765.}
  }
}
\description{
This function will parse a dataframe and compute mortality based on the Surgical Risk Scale (SRS). To use the function, you will need to manipulate your dataframe to have columns with the structure detailed below.
}
\section{Converting to probability scale}{

  The function will produce SRS_mort values which are on the log-odds scale
  To convert to probabilities (0 to 1 scale), use \code{arm::invlogit()}. See: \code{\link[arm]{invlogit}}.
}

\section{References}{

  \itemize{
  \item Sutton R, Bann S, Brooks M, Sarin S. The surgical risk scale as an improved tool for risk-adjusted analysis in comparative surgical audit. Br J Surg. 2002 Jun 1;89(6):763–8. \url{http://onlinelibrary.wiley.com/doi/10.1046/j.1365-2168.2002.02080.x/abstract}.
  }
}

\examples{
\dontrun{
#Example of pre-processing to rename data variables to match expected column names
library(tidyverse)

test_data <- raw_data \%>\%
select(ASA = S03AsaPsClass,
        OpUrgency = S02OperativeUrgency,
        OpSeverity = S02PlannedProcSeverity)
}

test_data <- patients
test_output <- gen.SRS(test_data)
head(test_output)
}
