% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure_codes.R
\docType{data}
\name{procedure_codes}
\alias{procedure_codes}
\title{Procedure codes}
\format{A data frame with 1960 rows and 9 variables:
\describe{
  \item{SurgeryProcedureMainGroupId}{Integer corresponding Main Group ID used in SNAP-2}
  \item{SurgeryProcedureSubGroupId}{Integer corresponding Subgroup ID used in SNAP-2}
  \item{SurgeryProcedureId}{Integer corresponding procedure ID used in SNAP-2}
  \item{Code}{Integer corresponding procedure ID used in SNAP-2}
  \item{SurgeryProcedure}{Character string name and description of the procedure}
  \item{SurgeryProcedureSeverity}{Surgical severity of the procedure as per AXA-PPP}
  \item{MainGroup}{Character string name and description of the organ system corresponding to Main Group ID}
  \item{SubGroup}{Character string name and description of the Subgroup within the main organ system group of the procedure}
  \item{Specialty}{Character string name and description of the surgical specialty}

}}
\source{
The procedure severity is available from AXA-PPP \url{https://online.axappphealthcare.co.uk/SpecialistForms/SpecialistCode.mvc/Print?source=contracted}.
  Example of papers that have used these procedure severity codes: \url{https://doi.org/10.1093/bja/aex117}
}
\usage{
procedure_codes
}
\description{
A dataset containing the procedure codes used in SNAP-2 with the corresponding AXA-PPP
procedure severity (Minor, Intermediate, Major, Xmajor, Complex) and the specialty name.
}
\keyword{datasets}
